requirements.txt
flask==3.0.2
line-bot-sdk==3.17.1
google-generativeai==0.8.5
python-dotenv==1.0.1
pydantic>=2.0
werkzeug==3.0.2
typing-extensions>=4.5.0
